new clear architecture
