+++ go.mod
@@ -3,6 +3,7 @@ module main
 go 1.13
 
 require (
+	github.com/go-redis/redis v6.15.7+incompatible // indirect
 	github.com/go-redis/redis/v8 v8.0.0-alpha.1 // indirect
 	github.com/labstack/echo v3.3.10+incompatible // indirect
 	github.com/labstack/gommon v0.3.0 // indirect
