+++ internal/feeds/repository.go
@@ -1,0 +1,8 @@
+package feeds
+
+import "../models"
+
+type FeedRepository interface {
+	GetUserFeedById(int, int) (models.Feed, error)
+	GetUserFeedByEmail(string, int) (models.Feed, error)
+}
