+++ internal/feeds/usecase.go
@@ -1,0 +1,5 @@
+package feeds
+
+type FeedUseCase interface {
+	Feed(string) (map[string]interface{} , error)
+}
