+++ internal/feeds/usecase/feed_ucase.go
@@ -1,0 +1,40 @@
+package usecase
+
+import (
+	Ckie "../../cookies/repository"
+	SessRep "../../cookies"
+	"../../feeds"
+	FeedRep "../../feeds/repository"
+	"../../tools/errors"
+)
+
+type FeedUseCaseRealisation struct {
+	feedDB    feeds.FeedRepository
+	sessionDB SessRep.CookieRepository
+}
+
+func (feedR FeedUseCaseRealisation) Feed(cookie string) (map[string]interface{} , error) {
+
+	id, err := feedR.sessionDB.GetUserIdByCookie(cookie)
+
+	if err != nil {
+		return nil , errors.InvalidCookie
+	}
+
+	sendData := make(map[string]interface{})
+
+	sendData["feed"], err = feedR.feedDB.GetUserFeedById(id, 30)
+
+	if err != nil {
+		return nil , errors.FailReadFromDB
+	}
+
+	return sendData , nil
+}
+
+func NewFeedUseCaseRealisation(feedDB FeedRep.FeedRepositoryRealisation, sesDB Ckie.CookieRepositoryRealisation) FeedUseCaseRealisation {
+	return FeedUseCaseRealisation{
+		feedDB:   feedDB,
+		sessionDB: sesDB,
+	}
+}
