+++ internal/models/feed.go
@@ -1,0 +1,5 @@
+package models
+
+type Feed struct {
+	Posts []Post `json:"posts,omitempty"`
+}
