+++ internal/models/friends.go
@@ -1,0 +1,5 @@
+package models
+
+type Friends struct{
+	IsFriends  bool   `json:"isfriends,omitempty"`
+}
