+++ internal/models/photo.go
@@ -1,0 +1,7 @@
+package models
+
+type Photo struct {
+	Url     string `json:"url,omitempty"`
+	Likes   int    `json:"likes,omitempty"`
+	WasLike bool   `json:"wasLike,omitempty"`
+}
