+++ internal/tools/errors/errors_types.go
