--- middleware/middleware.go
@@ -1,79 +1,0 @@
-package middleware
-
-import (
-	"../models"
-	"github.com/labstack/echo"
-	uuid "github.com/satori/go.uuid"
-	"go.uber.org/zap"
-	"net/http"
-	"time"
-)
-
-func SetCorsMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
-
-	return func(c echo.Context) error {
-
-		//TODO: убрать из корса
-		//c.Response().Header().Set("Content-Type", "application/json; charset=utf-8")
-
-		c.Response().Header().Set("Access-Control-Allow-Origin", "http://localhost:3000")
-		c.Response().Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS, PUT, DELETE, POST")
-		c.Response().Header().Set("Set-Cookie", "HttpOnly, Secure, SameSite=Strict")
-		c.Response().Header().Set("Access-Control-Allow-Headers", "Origin, X-Login, Set-Cookie, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, csrf-token, Authorization")
-		c.Response().Header().Set("Access-Control-Allow-Credentials", "true")
-		c.Response().Header().Set("Vary", "Cookie")
-
-		if c.Request().Method == http.MethodOptions {
-			return c.NoContent(http.StatusOK)
-		}
-
-		return next(c)
-
-	}
-}
-
-func PanicMiddleWare(next echo.HandlerFunc) echo.HandlerFunc {
-
-	return func(c echo.Context) error {
-
-		defer func() error {
-			if err := recover(); err != nil {
-				return c.JSON(http.StatusInternalServerError, models.JsonStruct{Err: "server panic ! "})
-			}
-			return next(c)
-		}()
-		return next(c)
-
-	}
-}
-
-func AccessLog(logs *zap.SugaredLogger) echo.MiddlewareFunc {
-
-	return func(next echo.HandlerFunc) echo.HandlerFunc {
-		logger := logs
-
-		return func(rwContext echo.Context) error {
-
-			uniqueID, _ := uuid.NewV4()
-			start := time.Now()
-			rwContext.Response().Header().Set("REQUEST_ID", uniqueID.String())
-
-			logger.Info(
-				zap.String("ID", uniqueID.String()),
-				zap.String("URL", rwContext.Request().URL.Path),
-				zap.String("METHOD", rwContext.Request().Method),
-			)
-
-			err := next(rwContext)
-
-			logger.Info(
-				zap.String("ID", uniqueID.String()),
-				zap.Duration("TIME FOR ANSWER", time.Since(start)),
-			)
-
-			return err
-
-		}
-	}
-
-}
