--- models/feed.go
@@ -1,5 +1,0 @@
-package models
-
-type Feed struct {
-	Posts []Post `json:"posts,omitempty"`
-}
