--- models/friends.go
@@ -1,6 +1,0 @@
-package models
-
-
-type Friends struct{
-	IsFriends  bool   `json:"isfriends,omitempty"`
-}
