--- repository/cookie.go
@@ -1,9 +1,0 @@
-package repository
-
-import "time"
-
-type CookieRepository interface {
-	AddCookie(int, string, time.Duration) error
-	DeleteCookie(string) error
-	GetUserIdByCookie(string) (int, error)
-}
