--- repository/feed.go
@@ -1,8 +1,0 @@
-package repository
-
-import "../models"
-
-type FeedRepository interface {
-	GetUserFeedById(int, int) (models.Feed, error)
-	GetUserFeedByEmail(string, int) (models.Feed, error)
-}
