--- usecase/feed.go
@@ -1,5 +1,0 @@
-package usecase
-
-type FeedUseCase interface {
-	Feed(string) (map[string]interface{} , error)
-}
